pyproject.toml
[tool.poetry]
name = "lnproxy"
version = "0.1.0"
description = "Proxy connections from a patched C-Lightning"
authors = ["Will Clark"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
trio = "^0.13.0"
pylightning = "^0.0.7"
gotenna = {url = "https://github.com/gotenna/PublicSDK/raw/master/python-public-sdk/goTenna-0.12.5-py3-none-any.whl"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
